Pagina de inicio realizada

+++ Administrador/Marcas/ProcesoEliminarMarca.php
@@ -11,11 +11,27 @@
     // capturar tipo
     $tipo   = $_GET["TIPO"];
     // crear sentencia SQL para eliminar registro
-    $sql = "DELETE FROM marcas WHERE marcasID=$id";
-    // die($sql);
-    // ejecutar sentencia SQL
-    mysql_query($sql,$conex);
-    // redirigir al listado
-    header("Location: ../../index.php?ORD=$orden&TIPO=$tipo");   
+    $sql = "SELECT * FROM producto WHERE productoMARCA='$id'";
+    // ejecutar sentencia de control
+    $result = mysql_query($sql,$conex);
+    // determinar exitencia
+    if (mysql_num_rows($result)==0) {
+        // crear sentencia SQL para insertar
+        // crear sentencia SQL para eliminar registro
+        $sql = "DELETE FROM marcas WHERE marcasID=$id";
+        // die($sql);
+        // ejecutar sentencia SQL
+        mysql_query($sql,$conex);  
+        // cerrar conexión
+        mysql_close($conex);
+        // volver automáticamente al formulario (redirigir)
+        header("location: ../../DatosOK/CambioContraOK.php?MSG=Se borro correctamente la marca");        
+    } else {
+        // cerrar conexión
+        mysql_close($conex);        
+        // enviar mensaje de error
+        header("location:  ../../Error/errorPage.php?MSG=La marca tiene productos asociados.");   
+  
+    } // endif
 
 ?>
